align

+++ src/types/mod.rs
@@ -5,22 +5,20 @@ pub use constraints::{EdgeConstraint, GlobalConstraint, VertexConstraint, Vertex
 mod constraints;
 
 /// The vertex id type.
-pub type VId = i32;
+pub type VId = i64;
 
 /// The vertex label type.
-pub type VLabel = i16;
+pub type VLabel = i64;
 
 /// The edge label type.
-pub type ELabel = i16;
+pub type ELabel = i64;
 
-#[repr(C, packed)]
 pub(crate) struct VLabelPosLen {
     pub vlabel: VLabel,
     pub pos: usize,
     pub len: usize,
 }
 
-#[repr(C, packed)]
 pub(crate) struct VertexHeader {
     pub num_bytes: usize,
     pub vid: VId,
@@ -29,27 +27,23 @@ pub(crate) struct VertexHeader {
     pub num_vlabels: usize,
 }
 
-#[repr(C, packed)]
 pub(crate) struct NeighborHeader {
     pub nid: VId,
     pub num_n_to_v: usize,
     pub num_v_to_n: usize,
 }
 
-#[repr(C, packed)]
 pub(crate) struct SuperRowHeader {
     pub num_rows: usize,
     pub num_eqvs: usize,
     pub num_vertices: usize,
 }
 
-#[repr(C, packed)]
 pub(crate) struct PosLen {
     pub pos: usize,
     pub len: usize,
 }
 
-#[repr(C, packed)]
 pub(crate) struct VIdPos {
     pub vid: VId,
     pub pos: usize,
